+++ lib/td_dd_web/policy.ex
@@ -30,6 +30,9 @@ defmodule TdDdWeb.Policy do
   def authorize(:query, %{} = claims, :data_structure),
     do: Bodyguard.permit(TdDd.DataStructures, :query, claims)
 
+  def authorize(:query, %{} = claims, :functions),
+    do: Bodyguard.permit(TdDq.Functions, :query, claims)
+
   def authorize(:query, %{} = claims, :implementation),
     do: Bodyguard.permit(TdDq.Implementations, :query, claims)